Refactor code from LayoutTree
