+++ panda.config.ts
@@ -119,6 +119,17 @@ const keyframes = defineKeyframes({
       fill: 'white',
     },
   },
+  bobble: {
+    '0%': {
+      transform: 'translateX(-50%) translateY(0)',
+    },
+    '50%': {
+      transform: 'translateX(-50%) translateY(10px)',
+    },
+    '100%': {
+      transform: 'translateX(-50%) translateY(0)',
+    },
+  },
 })
 
 const globalCss = defineGlobalStyles({
